+++ Sources/StreamChatSwiftUI/ChatChannel/Composer/PhotoAttachmentPickerView.swift
@@ -102,7 +102,7 @@ public struct PhotoAttachmentCell: View {
                             .allowsHitTesting(true)
                             .onTapGesture {
                                 withAnimation {
-                                    if let assetURL = assetURL {
+                                    if let assetURL = asset.mediaType == .image ? assetJpgURL() : assetURL {
                                         onImageTap(
                                             AddedAsset(
                                                 image: image,
@@ -192,5 +192,14 @@ public struct PhotoAttachmentCell: View {
                 loading = false
             }
         }
+    }
+
+    /// The original photo is usually in HEIC format.
+    /// This makes sure that the photo is converted to JPG.
+    /// This way it is more compatible with other platforms.
+    private func assetJpgURL() -> URL? {
+        guard let assetURL = assetURL else { return nil }
+        guard let assetData = try? Data(contentsOf: assetURL) else { return nil }
+        return try? UIImage(data: assetData)?.saveAsJpgToTemporaryUrl()
     }
 }