+++ Sources/StreamChatSwiftUI/ChatChannel/MessageList/Polls/PollAttachmentView.swift
@@ -55,6 +55,7 @@ public struct PollAttachmentView<Factory: ViewFactory>: View {
             ForEach(options.prefix(PollAttachmentViewModel.numberOfVisibleOptionsShown)) { option in
                 PollOptionView(
                     viewModel: viewModel,
+                    factory: factory,
                     option: option,
                     optionVotes: poll.voteCount(for: option),
                     maxVotes: poll.currentMaximumVoteCount,
@@ -73,7 +74,7 @@ public struct PollAttachmentView<Factory: ViewFactory>: View {
                     )
                 }
                 .fullScreenCover(isPresented: $viewModel.allOptionsShown) {
-                    PollAllOptionsView(viewModel: viewModel)
+                    PollAllOptionsView(viewModel: viewModel, factory: factory)
                 }
             }
             
@@ -115,7 +116,7 @@ public struct PollAttachmentView<Factory: ViewFactory>: View {
                     Text(L10n.Message.Polls.Button.viewNumberOfComments(viewModel.poll.answersCount))
                 }
                 .fullScreenCover(isPresented: $viewModel.allCommentsShown) {
-                    PollCommentsView(poll: viewModel.poll, pollController: viewModel.pollController)
+                    PollCommentsView(factory: factory, poll: viewModel.poll, pollController: viewModel.pollController)
                 }
             }
             
@@ -125,7 +126,7 @@ public struct PollAttachmentView<Factory: ViewFactory>: View {
                 Text(L10n.Message.Polls.Button.viewResults)
             }
             .fullScreenCover(isPresented: $viewModel.pollResultsShown) {
-                PollResultsView(viewModel: viewModel)
+                PollResultsView(viewModel: viewModel, factory: factory)
             }
             
             if viewModel.showEndVoteButton {
@@ -182,10 +183,11 @@ public struct PollAttachmentView<Factory: ViewFactory>: View {
 
 extension PollOption: Identifiable {}
 
-struct PollOptionView: View {
+struct PollOptionView<Factory: ViewFactory>: View {
     
     @ObservedObject var viewModel: PollAttachmentViewModel
     
+    let factory: Factory
     let option: PollOption
     var optionFont: Font = InjectedValues[\.fonts].body
     var optionVotes: Int?
@@ -221,9 +223,13 @@ struct PollOptionView: View {
                             ForEach(
                                 option.latestVotes.prefix(2)
                             ) { vote in
-                                MessageAvatarView(
-                                    avatarURL: vote.user?.imageURL,
-                                    size: .init(width: 20, height: 20)
+                                factory.makeMessageAvatarView(
+                                    for: UserDisplayInfo(
+                                        id: vote.user?.id ?? "",
+                                        name: vote.user?.name ?? "",
+                                        imageURL: vote.user?.imageURL,
+                                        size: .init(width: 20, height: 20)
+                                    )
                                 )
                             }
                         }