+++ Sources/StreamChatSwiftUI/Utils/Common/UIImage+Extensions.swift
@@ -42,7 +42,7 @@ extension UIImage {
 }
 
 extension UIImage {
-    func temporaryLocalFileUrl() throws -> URL? {
+    func saveAsJpgToTemporaryUrl() throws -> URL? {
         guard let imageData = jpegData(compressionQuality: 1.0) else { return nil }
         let imageName = "\(UUID().uuidString).jpg"
         let documentDirectory = NSTemporaryDirectory()